Move tedge-write tests to RobotFramework
To properly test tedge-write, we need to be running in the environment
where

1. sudo is installed
2. we run as `tedge` user
3. there is a sudoers entry that allows tedge to `sudo tedge-write`
   without password

RobotFramework provides that environment, so tests of tedge-write
functionality were moved there.
