+++ crates/core/tedge_write/Cargo.toml
@@ -22,7 +22,6 @@ uzers.workspace = true
 which.workspace = true
 
 [dev-dependencies]
-assert_cmd.workspace = true
 tempfile.workspace = true
 
 [lints]
